Refactor and various updates.
- Update README.md.
- Move configs into `configs/`.
- Add "recommended" configs to cover common use cases.
- Update templates.
