+++ README.md
@@ -2,15 +2,26 @@
 
 This package provides eslint rules used by Digital Bazaar as a set of extendable shared configs.
 
-There are 6 rule sets:
-1. `eslint-config-digitalbazaar`: Base rules for both node and the browser.
-2. `eslint-config-digitalbazaar/import`: Rules for es6 imports in the browser and node.
-3. `eslint-config-digitalbazaar/jsdoc`: Rules for JSDoc for both node and the browser.
-4. `eslint-config-digitalbazaar/module`: Rules for modules for both node and the browser.
-5. `eslint-config-digitalbazaar/vue2`: Rules for Vue 2 projects and browser only.
-6. `eslint-config-digitalbazaar/vue3`: Rules for Vue 3 projects and browser only.
+There are various core rule sets:
+- `eslint-config-digitalbazaar`: Base rules for both node and browser code.
+- `eslint-config-digitalbazaar/import`: Rules for es6 imports in the browser and node code.
+- `eslint-config-digitalbazaar/jsdoc`: Rules for JSDoc for both node and browser code.
+- `eslint-config-digitalbazaar/module`: Rules for modules for both node and browser code.
+- `eslint-config-digitalbazaar/vue2`: Rules for Vue 2 code.
+- `eslint-config-digitalbazaar/vue3`: Rules for Vue 3 code.
+
+For common use cases, there are recommended configs, that load the core rules:
+- `eslint-config-digitalbazaar/browser-recommended`: Recommended rules for projects targeting web browsers.
+- `eslint-config-digitalbazaar/node-recommended`: Recommended rules for projects targeting node.
+- `eslint-config-digitalbazaar/recommended`: Common recommended rules.
+- `eslint-config-digitalbazaar/universal-recommended`: Recommended rules for projects targeting node and web browsers.
+- `eslint-config-digitalbazaar/vue2-recommended`: Recommended rules for projects targeting Vue 2.
+- `eslint-config-digitalbazaar/vue3-recommended`: Recommended rules for projects targeting Vue 3.
 
 ## Installation
+
+Requires:
+- eslint v9
 
 If you do not have eslint installed:
 ```
@@ -24,28 +35,33 @@ npx eslint --init
 ```
 or rename a template from the templates dir
 ```
-cp node_modules/eslint-config-digitalbazaar/templates/node.js ./eslint.config.js
+cp node_modules/eslint-config-digitalbazaar/templates/node-recommended.js ./eslint.config.js
 ```
 
 ## Usage
 
-All rules can be accessed via the shorthand `digitalbazaar`
-or using the full module name `eslint-config-digitalbazaar`.
-
-ESLint's documentation on [shareable configs](https://eslint.org/docs/developer-guide/shareable-configs) can be found here.
+Config files can use the simple array syntax or the `defineConfig` syntax
+(see [shareable configs](https://eslint.org/docs/developer-guide/shareable-configs)).
 
 Example `eslint.config.js` root setup:
 ```js
-module.exports = {
-  root: true,
-  // using full module name
-  extends: ['eslint-config-digitalbazaar']
-}
+import config from 'eslint-config-digitalbazaar/recommended'
+
+export default [
+  ...config
+];
 ```
+
+Recommended configs (listed above) exist for common use cases. They may be
+sufficient on their own when using the opinionated code and file style used by
+Digital Bazaar for specific types of pacakges. However, they can also be mixed
+and matched and modified as needed. Note that currently required plugin
+dependencies must be manually installed when using recommended configs.
 
 ### Imports
 
-To use the import rules you need to install [`eslint-plugin-import`](https://github.com/import-js/eslint-plugin-import):
+To use the `import` plugin rules you need to install
+[`eslint-plugin-import`](https://github.com/import-js/eslint-plugin-import):
 
 ```
 npm i -D eslint-plugin-import
@@ -53,39 +69,54 @@ npm i -D eslint-plugin-import
 
 Example `eslint.config.js` import setup:
 ```js
-module.exports = {
-  extends: ['digitalbazaar/import']
-}
+import config from 'eslint-config-digitalbazaar/recommended'
+import importConfig from 'eslint-config-digitalbazaar/import'
+
+export default [
+  ...config,
+  ...importConfig
+];
 ```
 
 ### JSDoc
 
-To use the JSDoc rules you will need to install [`eslint-plugin-jsdoc`](https://github.com/gajus/eslint-plugin-jsdoc):
+When using the recommended configs, or to use the JSDoc rules standalone, you
+need to install
+[`eslint-plugin-jsdoc`](https://github.com/gajus/eslint-plugin-jsdoc):
+
 ```
 npm i -D eslint-plugin-jsdoc
 ```
 
 Example `eslint.config.js` JSDoc setup:
 ```js
-module.exports = {
-  // only the JSDoc rules and any rules in parent dirs
-  extends: ['digitalbazaar/jsdoc']
-}
+import config from 'eslint-config-digitalbazaar'
+import jsdocConfig from 'eslint-config-digitalbazaar/jsdoc'
+
+export default [
+  ...config,
+  ...jsdocConfig
+];
 ```
 
 ### Modules
 
-To use ES module code rather than CommonJS, you will need to install [`eslint-plugin-unicorn`](https://github.com/sindresorhus/eslint-plugin-unicorn):
+When using the recommended configs, or to use the ES module rules standalone,
+you need to install
+[`eslint-plugin-unicorn`](https://github.com/sindresorhus/eslint-plugin-unicorn):
 ```
 npm i -D eslint-plugin-unicorn
 ```
 
 Example `eslint.config.js` ESM setup:
 ```js
-module.exports = {
-  // only the module rules and any rules in parent dirs
-  extends: ['digitalbazaar/module']
-}
+import config from 'eslint-config-digitalbazaar'
+import moduleConfig from 'eslint-config-digitalbazaar/module'
+
+export default [
+  ...config,
+  ...moduleConfig
+];
 ```
 
 ### Vue 2
@@ -97,10 +128,13 @@ npm i -D eslint-plugin-vue
 
 Example `eslint.config.js` Vue setup:
 ```js
-module.exports = {
-  // only the vue rules and any rules in parent dirs
-  extends: ['digitalbazaar/vue2']
-}
+import config from 'eslint-config-digitalbazaar'
+import vue2Config from 'eslint-config-digitalbazaar/vue2'
+
+export default [
+  ...config,
+  ...vue2Config
+];
 ```
 
 For command line use you may need to [explicitly enable linting `.vue`
@@ -115,10 +149,13 @@ npm i -D eslint-plugin-vue
 
 Example `eslint.config.js` Vue setup:
 ```js
-module.exports = {
-  // only the vue3 rules and any rules in parent dirs
-  extends: ['digitalbazaar/vue3']
-}
+import config from 'eslint-config-myconfig'
+import vue3Config from 'eslint-config-digitalbazaar/vue3'
+
+export default [
+  ...config,
+  ...vue3Config
+];
 ```
 
 For command line use you may need to [explicitly enable linting `.vue`
@@ -126,19 +163,18 @@ files](https://eslint.vuejs.org/user-guide/#running-eslint-from-the-command-line
 
 ### Composition
 
-The rules do not depend on each other and are composable:
+The core rules do not depend on each other and are composable:
 ```js
-module.exports = {
-  extends: [
-    'digitalbazaar',
-    'digitalbazaar/jsdoc',
-    'digitalbazaar/module'
-    'digitalbazaar/vue3'
-  ] // 4 rule sets in one file using shorthand.
-}
-```
+import config from 'eslint-config-digitalbazaar'
+import jsdocConfig from 'eslint-config-digitalbazaar/jsdoc'
+import vue3Config from 'eslint-config-digitalbazaar/vue3'
 
-The rules can also be used together via [cascade configuration](https://eslint.org/docs/user-guide/configuring).
+export default [
+  ...config,
+  ...jsdoconfig
+  ...vue3Config
+];
+```
 
 ## Other Rules
 
@@ -146,7 +182,8 @@ Other rules that are not included above but can be useful:
 
 ### `unicorn/prefer-node-protocol`
 
-Use `node:module` style for Node.js modules.
+- Use `node:module` style for Node.js modules.
+- This is included in the node related recommended configs.
 
 See [`unicorn/prefer-node-protocol`](https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-node-protocol.md).
 
@@ -157,7 +194,14 @@ npm i -D eslint-plugin-unicorn
 
 Rules:
 ```js
-  rules: {
-    'unicorn/prefer-node-protocol': 'error'
+import config from 'eslint-config-digitalbazaar'
+
+export default [
+  ...config,
+  {
+    rules: {
+      'unicorn/prefer-node-protocol': 'error'
+    }
   }
+];
 ```
