+++ configs/base.js
