+++ configs/browser-recommended.js
@@ -1,0 +1,15 @@
+import globals from 'globals';
+
+import recommended from './recommended.js';
+
+export default [
+  ...recommended,
+  {
+    languageOptions: {
+      globals: {
+        ...globals.browser
+      }
+    }
+  }
+];
+
