+++ configs/module.js
