+++ configs/node-recommended.js
@@ -1,0 +1,20 @@
+import globals from 'globals';
+
+import recommended from './recommended.js';
+
+export default [
+  ...recommended,
+  {
+    languageOptions: {
+      globals: {
+        ...globals.node
+      }
+    }
+  },
+  {
+    // misc rules
+    rules: {
+      'unicorn/prefer-node-protocol': 'error'
+    }
+  }
+];
