+++ configs/universal-recommended.js
@@ -1,0 +1,7 @@
+import browser from './browser-recommended.js';
+import node from './node-recommended.js';
+
+export default [
+  ...node,
+  ...browser
+];
