+++ configs/vue-common.js
