+++ configs/vue2-recommended.js
@@ -1,0 +1,7 @@
+import browser from './browser-recommended.js';
+import vue2 from './vue2.js';
+
+export default [
+  ...browser,
+  ...vue2
+];
