+++ configs/vue2.js
