+++ configs/vue3-recommended.js
@@ -1,0 +1,7 @@
+import browser from './browser-recommended.js';
+import vue3 from './vue3.js';
+
+export default [
+  ...browser,
+  ...vue3
+];
