+++ configs/vue3.js
