+++ templates/browser.js
@@ -1,0 +1,5 @@
+import config from 'eslint-config-digitalbazaar/browser-recommended';
+
+export default [
+  ...config
+];
