+++ templates/node.js
@@ -1,13 +1,5 @@
-import digitalbazaar from 'eslint-config-digitalbazaar';
-import globals from 'globals';
+import config from 'eslint-config-digitalbazaar/node-recommended';
 
 export default [
-  {
-    languageOptions: {
-      globals: {
-        ...globals.node
-      }
-    }
-  },
-  ...digitalbazaar
+  ...config
 ];
