+++ templates/universal.js
@@ -1,0 +1,5 @@
+import config from 'eslint-config-digitalbazaar/universal-recommended';
+
+export default [
+  ...config
+];
